Fixing a bug. Previous version: After coadding spectra by using combspec in coadd1d.py, wavelengths were unexpectedly shifted. New version: Wavelengths are not artificially shifted (as it should be). Reason for the bug: Precision in weights_flat in compute_stack was not enough high. Modification: Weights_flat set to float64 type (instead of float32).

+++ pypeit/core/coadd1d.py
@@ -1273,7 +1273,7 @@ def compute_stack(wave_grid, waves, fluxes, ivars, masks, weights):
     fluxes_flat = fluxes[ubermask].flatten()
     ivars_flat = ivars[ubermask].flatten()
     vars_flat = utils.inverse(ivars_flat)
-    weights_flat = weights[ubermask].flatten()
+    weights_flat = np.float64(weights[ubermask].flatten())
 
     # Counts how many pixels in each wavelength bin
     nused, wave_edges = np.histogram(waves_flat,bins=wave_grid,density=False)
